Cleanup of structure
